Move to separate benchmark file; remove pointer test; remove dead comment
