+++ api/key/key_test.go
@@ -15,17 +15,13 @@
 package key_test
 
 import (
-	"context"
 	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 
 	"go.opentelemetry.io/otel/api/core"
-	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/api/key"
-	"go.opentelemetry.io/otel/api/trace"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
 func TestKeyValueConstructors(t *testing.T) {
@@ -205,62 +201,4 @@ func TestInfer(t *testing.T) {
 			t.Errorf("+got, -want: %s", diff)
 		}
 	}
-}
-
-// Note: The tests below load a real SDK
-func getSpan() trace.Span {
-	_, sp := global.Tracer("Test").Start(context.Background(), "Span")
-	tr, _ := sdktrace.NewProvider()
-	global.SetTraceProvider(tr)
-	return sp
-}
-
-func BenchmarkMultiNoKeyInference(b *testing.B) {
-	sp := getSpan()
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		sp.SetAttributes(key.Int("Attr", 1))
-	}
-}
-
-func BenchmarkMultiWithKeyInference(b *testing.B) {
-	sp := getSpan()
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		sp.SetAttributes(key.Infer("Attr", 1))
-	}
-}
-
-func BenchmarkSingleWithKeyInferenceValue(b *testing.B) {
-	sp := getSpan()
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		sp.SetAttribute("Attr", 1)
-	}
-
-	b.StopTimer()
-}
-
-func BenchmarkSingleWithKeyInferencePointer(b *testing.B) {
-	sp := getSpan()
-	x := new(int)
-	*x = 1
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		sp.SetAttribute("Attr", x)
-	}
-
-	b.StopTimer()
 }
